+++ src/main/java/model/board/element/powerup/benefit/MaxBombsBonus.java
@@ -26,7 +26,7 @@ public class MaxBombsBonus extends Bonus {
      * @param visible the visibility status of the bonus
      * @param owner   the player who owns the bonus
      */
-    public MaxBombsBonus(int x, int y, int width, int height, double velocity, Image image, boolean alive, boolean visible, Player owner) {
+    public MaxBombsBonus(double x, double y, int width, int height, double velocity, Image image, boolean alive, boolean visible, Player owner) {
         super(x, y, width, height, velocity, image, alive, visible, owner);
     }
